use globals of the function; don't support local ctx; add tests in separated modules.
